Make Recap configurable
Use dynaconf to Recap. With Dynaconf users can configure Recap using environment
variables or config files. It also supports some snazzy features like config
validation and environments, which I'm not using right now.

As part of this change, I've added module context managers for the storage and
crawler modules. I'm don't totally know what I'm doing with Python, but this
approach seems to be a simple way to make the crawlers and storage layer
pluggable and configurable via configs.

I've added support for three very basic blocks:

* api
* crawlers
* storage

You can, of course, enable a bunch of Dynaconf stuff via environment variables;
things like environments (ENVIRONMENTS_FOR_DYNACONF, ENVVAR_PREFIX_FOR_DYNACONF,
etc), but that'll be described in Recap's documentation.
