+++ pyproject.toml
@@ -11,9 +11,10 @@ dependencies = [
     "fsspec>=2022.11.0",
     "httpx>=0.23.1",
     "typer>=0.7.0",
-    "tomli>=2.0.1",
+    # TODO should we move this into recap[db] or something?
     "psycopg2>=2.9.5",
     "sqlalchemy>=1.4.45",
+    "dynaconf>=3.1.11",
 ]
 requires-python = ">=3.10"
 readme = "README.md"
