+++ recap/cli.py
@@ -1,4 +1,7 @@
 import typer
+from recap import storage
+from recap.config import settings
+
 
 app = typer.Typer()
 
@@ -6,30 +9,25 @@ app = typer.Typer()
 @app.command()
 def api():
     import uvicorn
-    uvicorn.run(
-        "recap.api:app",
-        host="0.0.0.0",
-        port=8000,
-    )
+
+    with storage.open(**settings['storage']) as s:
+        # TODO user storage here
+        uvicorn.run(
+            "recap.api:app",
+            host=settings('api.host', '0.0.0.0'),
+            port=settings('api.port', 8000, cast=int),
+        )
 
 
 @app.command()
-def crawler(recap_url, crawl_url):
-    import httpx
-    import sqlalchemy as sa
-    from .crawler.db import Instance, Crawler
-    from .storage import RecapTableStorage
-    with httpx.Client(base_url=recap_url) as client:
-        # TODO read crawler configs from a TOML file
-        engine = sa.create_engine(crawl_url)
-        instance = Instance(engine)
-        storage = RecapTableStorage(
-            'postgresql',
-            'sticker_space_dev',
-            client
-        )
-        crawler = Crawler(storage, instance)
-        crawler.crawl()
+def crawler():
+    from recap import crawler
+
+    with storage.open(**settings['storage']) as s:
+        for infra, instance_dict in settings['crawlers'].items():
+            for instance, instance_config in instance_dict.items():
+                with crawler.open(s, **instance_config) as c:
+                    c.crawl()
 
 
 if __name__ == "__main__":
